@@ -11,50 +11,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "@ng-bootstrap/ng-bootstrap": "^3.3.1",
-    "bootstrap": "^4.1.3",
+    "@angular/animations": "^7.2.11",
+    "@angular/common": "^7.2.11",
+    "@angular/compiler": "^7.2.11",
+    "@angular/core": "^7.2.11",
+    "@angular/forms": "^7.2.11",
+    "@angular/http": "^7.2.11",
+    "@angular/platform-browser": "^7.2.11",
+    "@angular/platform-browser-dynamic": "^7.2.11",
+    "@angular/router": "^7.2.11",
+    "@ng-bootstrap/ng-bootstrap": "^4.1.0",
+    "bootstrap": "^4.3.1",
     "compass-mixins": "^0.12.10",
-    "core-js": "^2.5.4",
+    "core-js": "^3.0.0",
     "font-awesome": "^4.7.0",
-    "i": "^0.3.6",
     "jwt-decode": "^2.2.0",
+    "material-design-icons": "^3.0.1",
     "mdi": "^2.2.43",
     "moment": "^2.24.0",
-    "npm": "^6.9.0",
-    "rxjs": "^6.2.2",
+    "rxjs": "^6.4.0",
     "rxjs-compat": "^6.4.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.6",
-    "@angular/cli": "~6.2.5",
-    "@angular/compiler-cli": "^7.2.9",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
-    "file-loader": "^2.0.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "^0.13.7",
+    "@angular/cli": "^7.3.7",
+    "@angular/compiler-cli": "^7.2.11",
+    "@angular/language-service": "^7.2.11",
+    "@types/jasmine": "~3.3.12",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~11.12.2",
+    "codelyzer": "~5.0.0",
+    "file-loader": "^3.0.1",
+    "i": "0.3.6",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.0.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "npm": "6.9.0",
+    "protractor": "~5.4.2",
     "sass-loader": "^7.1.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "^3.2.4"
+    "ts-node": "~8.0.3",
+    "tslint": "~5.14.0",
+    "typescript": "3.2.1"
   }
 }
