@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {AuthenticationService} from "../services/authentication.service";
+import { FormBuilder, FormGroup, Validators} from '@angular/forms';
+import { AuthenticationService} from "../_services/authentication.service";
+import { first } from 'rxjs/operators';
+import { Router } from "@angular/router";
 @Component({
   selector: 'app-login-form',
   templateUrl: './login-form.component.html',
@@ -8,11 +10,13 @@ import {AuthenticationService} from "../services/authentication.service";
 })
 export class LoginFormComponent implements OnInit {
   loginForm: FormGroup;
-  constructor(private formBuilder: FormBuilder, private authService: AuthenticationService) { }
+  constructor(private formBuilder: FormBuilder, private authService: AuthenticationService, private router :Router) { }
 
   ngOnInit() {
     this.generateForm();
-    this.authService.logout();
+    if(this.authService.isLoggedIn()) {
+      this.router.navigate(['dashboard']);
+    }
   }
   // create form by using form builder
   private generateForm(): void {
@@ -26,15 +30,13 @@ export class LoginFormComponent implements OnInit {
     if(this.loginForm.invalid) {
       return;
     }
-    const username = this.loginForm.controls['username'].value;
-    const password = this.loginForm.controls['password'].value;
-    this.authService.login(username, password).subscribe(res => {
-    console.log(res);
-    this.authService.setSession(res);
-    console.log(this.authService.getExpiration());
-    console.log(this.authService.isLoggedIn());
-
-    }, error1 => console.log(error1));
+    const username = this.getControl.username.value;
+    const password = this.getControl.password.value;
+    this.authService.login(username, password).pipe(first()).subscribe(res => {
+      this.router.navigate(['dashboard']);
+    }, errorLogin => {
+      console.log('Login form' + JSON.stringify(errorLogin));
+    } );
   }
 
   // get all controls of loginForm, can use this function with command like getControl.controlName;
